Refactors storage error to include stack trace (#1076)
## Describe your changes and why you are making these changes
This PR lays out the ground work for removing all of the `var Err... =
errors.New(...)` code pattern in our codebase, as this was preventing us
from capturing the right stack trace as part of the `DropboxError`.

This PR defines our own `errors` lib so that we don't need to import
several different `errors` packages in the same file. >80% of this
library is just calling the dropbox errors library. The only part that
is custom is checking for error equality. This has to be custom because
we will always have a mix of dropbox errors and standard errors in our
codebase. Consider something like `io.ErrNoFile`, which is defined in
the standard `io` package. If we want to perform equality correctly for
all errors we need to deal with cases where we could be dealing with
either a standard error or a dropbox error.

Checking for equality with dropbox errors requires us to ignore the
stack trace, because the stack trace will be different, and instead just
look at the outermost error message.

A follow on PR will address removing all of the other `var Err == `
patterns in our codebase and deal with changing the package import from
the dropbox errors pkg to our internal one.

## Related issue number (if any)
ENG 2580

## Loom demo (if any)

## Checklist before requesting a review
- [ ] I have created a descriptive PR title. The PR title should
complete the sentence "This PR...".
- [ ] I have performed a self-review of my code.
- [ ] I have included a small demo of the changes. For the UI, this
would be a screenshot or a Loom video.
- [ ] If this is a new feature, I have added unit tests and integration
tests.
- [ ] I have run the integration tests locally and they are passing.
- [ ] I have run the linter script locally (See `python3
scripts/run_linters.py -h` for usage).
- [ ] All features on the UI continue to work correctly.
- [ ] Added one of the following CI labels:
    - `run_integration_test`: Runs integration tests
- `skip_integration_test`: Skips integration tests (Should be used when
changes are ONLY documentation/UI)